requirements.txt
pytesseract
streamlit
google-generativeai
python-dotenv
Pillow